refactoring nome dei metodi
