+++ src/main/java/com/example/ingsoft/Controllers/Validation/NumberTFValidator.java
@@ -4,7 +4,7 @@ import javafx.scene.control.TextField;
 
 import java.util.*;
 
-public class NumberTFValidator implements  Validate{
+class NumberTFValidator implements  Validate{
     private final HashMap<TextField, Integer> numberToLength;
     private final HashMap<TextField,Boolean> textFieldEmpty;
     private TextField textField;
