+++ src/main/java/com/example/ingsoft/Controllers/Validation/Validate.java
@@ -1,7 +1,7 @@
 package com.example.ingsoft.Controllers.Validation;
 
 
-public interface Validate {
+interface Validate {
     final String cssRedBorder = "-fx-border-color: red ; -fx-border-width: 2px ;";
     public boolean validate();
 }
